+++ examples/next/with-emailpassword/pages/api/auth/[[...path]].tsx
@@ -10,7 +10,10 @@ export default async function superTokens(req: NextApiRequest, res: NextApiRespo
     await superTokensNextWrapper(
         async (next) => {
             // This is needed for production deployments with Vercel
-            res.setHeader("Cache-Control", "no-cache, no-store, max-age=0, must-revalidate");
+            if (!response.hasHeader("Cache-Control")) {
+                res.setHeader("Cache-Control", "no-cache, no-store, max-age=0, must-revalidate");
+            }
+
             await middleware()(req, res, next);
         },
         req,